# Splitting the loop into two separated loops.